BMP280 V0.2 working Temperature

+++ examples/i2c_BMP280/i2c_BMP280.ino
@@ -28,7 +28,8 @@ void loop()
     float pascal;
     bmp280.getPressure(pascal);
 
-    float temperature;
+    //float temperature;
+    int32_t temperature;
     bmp280.getTemperature(temperature);
     bmp280.triggerMeasurement();
